requirements.txt
streamlit
torch
transformers
opencv-python
pillow
matplotlib
requests
